Ajustando enum para aplicação
