+++ app/Domains/Animal/Entities/AnimalEntryDataEntity.php
@@ -2,11 +2,17 @@
 
 namespace App\Domains\Animal\Entities;
 
+use App\Casts\EnumCast;
+use App\Domains\Enums\OriginEnum;
 use App\Models\AnimalEntryData;
 
 class AnimalEntryDataEntity extends AnimalEntryData
 {
     protected $table = 'animal_entry_datas';
+
+    protected $casts = [
+        'origin' => EnumCast::class.':'.OriginEnum::class,
+    ];
 
     protected $fillable = [
         'animal_id',
