+++ app/Domains/User/Entities/UserStatusEntity.php
@@ -3,7 +3,7 @@
 namespace App\Domains\User\Entities;
 
 use App\Casts\EnumCast;
-use App\Enums\UserStatusEnum;
+use App\Domains\Enums\UserStatusEnum;
 use App\Models\User;
 use App\Models\UserStatus;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
