+++ app/Http/Middleware/EnsureClientTypeHeader.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Middleware;
 
+use App\Domains\Enums\OriginEnum;
 use Closure;
 use Illuminate\Http\Request;
 
@@ -11,7 +12,7 @@ class EnsureClientTypeHeader
     {
         $clientType = $request->header('X-Client-Type');
 
-        if (! $clientType || ! in_array($clientType, ['web', 'mobile'])) {
+        if (! $clientType || ! in_array($clientType, OriginEnum::values())) {
             return response()->json([
                 'message' => 'Cabeçalho X-Client-Type ausente ou inválido.',
             ], 400);
