+++ app/Http/Requests/Vetenirarian/UpdateVetenirarianRequest.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Requests;
 
-use App\Enums\LinkedTypeEnum;
-use App\Enums\UserStatusEnum;
+use App\Domains\Enums\LinkedTypeEnum;
+use App\Domains\Enums\UserStatusEnum;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
